@@ -1,5 +1,5 @@
 import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators } from '@angular/forms';
 import { Toastr } from 'src/app/services/toastr.service';
 import { Router, ActivatedRoute } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
@@ -16,7 +16,7 @@ export class EditFaqComponent implements OnInit {
   answerText:string='';
 
   public FAQId: any;
-  public FAQForm: FormGroup;
+  public FAQForm: UntypedFormGroup;
   public FAQDetailsObj: any;
   public FAQgroup:any=[];
 
@@ -33,11 +33,11 @@ export class EditFaqComponent implements OnInit {
 
     this.editor = new Editor();
     this.FAQId = this.activeRoute.snapshot.paramMap.get('id');
-    this.FAQForm = new FormGroup({
-      'group_id': new FormControl([], [Validators.required]),
-      'question': new FormControl(null, [Validators.required]),
-      'answer': new FormControl(null, [Validators.required]),
-      'is_active': new FormControl(true)
+    this.FAQForm = new UntypedFormGroup({
+      'group_id': new UntypedFormControl([], [Validators.required]),
+      'question': new UntypedFormControl(null, [Validators.required]),
+      'answer': new UntypedFormControl(null, [Validators.required]),
+      'is_active': new UntypedFormControl(true)
     });
   }
 
