@@ -1,5 +1,5 @@
 import { EventEmitter, Component, OnInit, Output } from '@angular/core';
-import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, UntypedFormBuilder } from '@angular/forms';
 import { Helper } from 'src/app/services/helper.service';
 import { BsModalRef } from 'ngx-bootstrap/modal'
 import { ManufacturerAddEditModalService } from './manufacturer-add-edit-modal.service';
@@ -13,21 +13,21 @@ import { HttpClient } from '@angular/common/http';
 export class ManufacturerAddEditModalComponent implements OnInit {
   @Output() onEventCompleted: EventEmitter<any> = new EventEmitter();
   modalEvent: any;
-  manufacturerForm: FormGroup;
+  manufacturerForm: UntypedFormGroup;
 
   constructor(
     private _helper:Helper,
     private http: HttpClient,
     private _bsModalRef:BsModalRef,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private _manufacturerAddEditModalService: ManufacturerAddEditModalService){
     this.manufacturerForm = this.formBuilder.group({
-      'id':new FormControl(null),
-      'name': new FormControl(null, [Validators.required]),
-      'address':new FormControl(null,[]),
-      'contact_number':new FormControl(null,[]),
-      'country':new FormControl('India',[]),
-      'is_active':new FormControl(null)
+      'id':new UntypedFormControl(null),
+      'name': new UntypedFormControl(null, [Validators.required]),
+      'address':new UntypedFormControl(null,[]),
+      'contact_number':new UntypedFormControl(null,[]),
+      'country':new UntypedFormControl('India',[]),
+      'is_active':new UntypedFormControl(null)
     });
   }
 
